created an images folder added images worked on a model
