--- server/resolvers/typeDefs.js
@@ -1,28 +1,0 @@
-const typeDefs = ` type User {
-          _id: ID
-          username: String
-          email: String
-          password: String
-        savedCars: [Car]
-        }
-  
-  type Query {
-    User: User
-  }
-
-
-  }
-    type Auth{
-      token:ID
-      user: User
-  
-      
-      }`
-
-     
-
-
-
-
-
-module.exports = typeDefs;
