+++ server/schemas/typeDefs.js
@@ -1,0 +1,32 @@
+const typeDefs = ` type User {
+          _id: ID
+          username: String
+          email: String
+          password: String
+        savedCars: [Car]
+        }
+  
+
+        type Car{
+        _id: ID
+        carMake: String
+        carModel:String
+        carYear: String
+        carType: String
+        Image: String
+        
+        }
+  type Query {
+    
+  users: [User]
+    user(username: String!): User
+
+  }
+    type Auth{
+      token:ID!
+      user: User
+  
+      
+      }`;
+
+module.exports = typeDefs;
